requirements.txt
pillow==11.0.0
setuptools==75.1.0
wheel==0.44.0
